+++ src/managers/RevocationManager.ts
@@ -23,14 +23,14 @@ export default class RevocationManager extends BaseManager<Revocation> {
 	async fetchRevocations(playername: string, communityId: string, cache = true): Promise<Revocation[]> {
 		const revocations = await fetch(`${this.apiurl}/revocations/community/${strictUriEncode(playername)}/${strictUriEncode(communityId)}`)
 			.then(r=>r.json())
-		if (!revocations || !revocations[0]) return null
+		if (!revocations || !revocations[0]) return []
 		if (cache) revocations.forEach((revocation: Revocation) => this.add(revocation))
 		return revocations
 	}
 	async fetchAllRevocations(playername: string, cache = true): Promise<Revocation[]> {
 		const revocations = await fetch(`${this.apiurl}/revocations/player/${strictUriEncode(playername)}`)
 			.then(r=>r.json())
-		if (!revocations || !revocations[0]) return null
+		if (!revocations || !revocations[0]) return []
 		if (cache) revocations.forEach((revocation: Revocation) => this.add(revocation))
 		return revocations
 	}